+++ src/components.d.ts
@@ -5,7 +5,7 @@
  * It contains typing information for all components that exist in this project.
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
-import { ChipType, ComboboxItem, DropdownItem, InputChangeEventDetail, PdButtonColor, PdButtonSize, PdButtonType, PdColumn, PdIconLocation, PdModalConfig, PdPagingLocation, PdPlacement, PdStatus, PdTableIconConfiguration, PdTableRow, PdTableStyle, SelectedEvent, SortDropdownItem, TabValue, TextFieldTypes, TextWrap } from "./interface";
+import { ChipType, ComboboxItem, DropdownItem, InputChangeEventDetail, PdButtonColor, PdButtonSize, PdButtonType, PdColumn, PdIconLocation, PdModalConfig, PdPagingLocation, PdPlacement, PdStatus, PdTableIconConfiguration, PdTableRow, PdTableStyle, SelectedEvent, SortDropdownItem, SortRevertItem, TabValue, TextFieldTypes, TextWrap } from "./interface";
 import { DateOption, Options } from "flatpickr/dist/types/options";
 export namespace Components {
     interface PdAlert {
@@ -890,6 +890,14 @@ export namespace Components {
           * Reset the selection of the dropdown
          */
         "reset": () => Promise<void>;
+        /**
+          * Enables the revert item at the bottom of the dropdown
+         */
+        "reverseItem": boolean;
+        /**
+          * Data used for the reverse item at the bottom of the dropdown
+         */
+        "reverseItemData": SortRevertItem;
         /**
           * Set a preselected entry by index
          */
@@ -2440,10 +2448,19 @@ declare namespace LocalJSX {
          */
         "label"?: string;
         "onPd-change"?: (event: PdSortCustomEvent<SortDropdownItem>) => void;
+        "onPd-revert"?: (event: PdSortCustomEvent<SortDropdownItem>) => void;
         /**
           * Placeholder when no item is selected
          */
         "placeholder"?: string;
+        /**
+          * Enables the revert item at the bottom of the dropdown
+         */
+        "reverseItem"?: boolean;
+        /**
+          * Data used for the reverse item at the bottom of the dropdown
+         */
+        "reverseItemData"?: SortRevertItem;
     }
     interface PdTable {
         /**