+++ src/components/pd-sort/pd-sort.stories.js
@@ -4,7 +4,7 @@ export default {
     title: 'Interactions/Sort',
     parameters: {
         actions: {
-            handles: ['pd-change'],
+            handles: ['pd-change', 'pd-revert'],
         },
         notes,
     },
@@ -12,6 +12,8 @@ export default {
         label: { controls: { type: 'string' } },
         items: { controls: { type: 'object' } },
         disabled: { controls: { type: 'boolean' } },
+        revertItem: { controls: { type: 'boolean' } },
+        revertItemData: { controls: { type: 'object' } },
         label: { controls: { type: 'string' } },
         placeholder: { controls: { type: 'string' } },
         itemCount: { controls: { type: 'number' } },
@@ -24,9 +26,11 @@ export default {
 
 export const Sort = (args) => {
     const pdSort = document.createElement('pd-sort');
+    pdSort.label = args.label;
     pdSort.items = args.items;
     pdSort.disabled = args.disabled;
-    pdSort.label = args.label;
+    pdSort.revertItem = args.revertItem;
+    pdSort.revertItemData = args.revertItemData;
     pdSort.placeholder = args.placeholder;
     pdSort.itemCount = args.itemCount;
     pdSort.emptyItem = args.emptyItem;
@@ -38,14 +42,15 @@ export const Sort = (args) => {
 
 Sort.args = {
     items: [
-        { id: '1', label: 'Name', value: 'name', sort: 'asc', selected: 'true' },
+        { id: '1', label: 'Name', value: 'name', sort: 'asc', selected: true },
         { id: '2', label: 'Dokumententitel', value: 'documentTitle', sort: 'desc' },
         { id: '3', label: 'Änderungsdatum', value: 'changedAt', sort: 'asc' },
         { id: '4', label: 'Erstellungsdatum', value: 'createdAt', sort: 'asc' },
         { id: '6', label: 'Author/in', value: 'author', sort: 'asc' },
-        { id: '7', label: 'Umkehren', value: 'reverse', bottomSection: 'true' },
     ],
     disabled: false,
+    revertItem: true,
+    revertItemData: { label: 'Sort. Umkehren', selected: false },
     label: 'Sortieren nach:',
     placeholder: '',
     itemCount: 7,