+++ changelog.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.6.16] - 2023-10-24
+
+### Changed
+
+-   #263 pd-sort:
+    -   the pd-change event emits the selected item (which has a sort property (asc | desc) --> sort the list that way )
+    -   the pd-revert event emits the selected item too (which has a sort property (asc | desc) --> sort the list THE OTHER WAY way than its saved in the sort property ❗)
+
 ## [2.6.15] - 2023-10-20
 
 ### Changed